tests/rest_db/requirements.txt
# Standalone Tavern API Testing Dependencies
tavern>=2.16.0
pytest>=7.0,<7.3
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.28.0
PyYAML>=6.0.0